requirements.txt
fastapi
uvicorn
openai
pillow
pytesseract
sentence-transformers
numpy
httpx
python-dotenv

